backend/pyproject.toml
[project]
name = "acutal-backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pydantic-settings>=2.6.0",
    "python-jose[cryptography]>=3.3.0",
    "httpx>=0.27.0",
    "bcrypt>=4.0.0",
    "email-validator>=2.0.0",
    "boto3>=1.35.0",
    "anthropic>=0.40.0",
    "pymupdf>=1.24.0",
    "python-docx>=1.1.0",
    "python-pptx>=1.0.0",
    "openpyxl>=3.1.0",
    "matplotlib>=3.9.0",
    "xlrd>=2.0.0",
    "python-multipart>=0.0.9",
    "reportlab>=4.0.0",
    "stripe>=8.0.0",
    "resend>=2.0.0",
    "jinja2>=3.1.0",
    "yt-dlp>=2024.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
]

[tool.setuptools.packages.find]
include = ["app*"]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"
